+++ src/Tests/Unit/Public/Invoke-AmazonImageModel.Tests.ps1
@@ -981,7 +981,7 @@ InModuleScope 'pwshBedrock' {
                 Mock -CommandName Invoke-BDRRModel {
                     $response
                     $Region         | Should -BeExactly 'us-west-2'
-                    $ModelID        | Should -BeExactly 'amazon.titan-image-generator-v1'
+                    $ModelID        | Should -BeExactly 'us.amazon.titan-image-generator-v1'
                     $AccessKey      | Should -BeExactly 'ak'
                     $SecretKey      | Should -BeExactly 'sk'
                     $ContentType    | Should -BeExactly 'application/json'
@@ -1005,7 +1005,7 @@ InModuleScope 'pwshBedrock' {
                 Mock -CommandName Invoke-BDRRModel {
                     $response
                     $Region             | Should -BeExactly 'us-west-2'
-                    $ModelID            | Should -BeExactly 'amazon.titan-image-generator-v1'
+                    $ModelID            | Should -BeExactly 'us.amazon.titan-image-generator-v1'
                     $Credential         | Should -Not -BeNullOrEmpty
                     $EndpointUrl        | Should -BeExactly 'string'
                     $NetworkCredential  | Should -Not -BeNullOrEmpty