+++ src/Tests/Unit/Public/Invoke-AnthropicModel.Tests.ps1
@@ -828,7 +828,7 @@ InModuleScope 'pwshBedrock' {
                 Mock -CommandName Invoke-BDRRModel {
                     $response
                     $Region         | Should -BeExactly 'us-west-2'
-                    $ModelID        | Should -BeExactly 'anthropic.claude-3-sonnet-20240229-v1:0'
+                    $ModelID        | Should -BeExactly 'us.anthropic.claude-3-sonnet-20240229-v1:0'
                     $AccessKey      | Should -BeExactly 'ak'
                     $SecretKey      | Should -BeExactly 'sk'
                     $ContentType    | Should -BeExactly 'application/json'
@@ -852,7 +852,7 @@ InModuleScope 'pwshBedrock' {
                 Mock -CommandName Invoke-BDRRModel {
                     $response
                     $Region             | Should -BeExactly 'us-west-2'
-                    $ModelID            | Should -BeExactly 'anthropic.claude-3-sonnet-20240229-v1:0'
+                    $ModelID            | Should -BeExactly 'us.anthropic.claude-3-sonnet-20240229-v1:0'
                     $Credential         | Should -Not -BeNullOrEmpty
                     $EndpointUrl        | Should -BeExactly 'string'
                     $NetworkCredential  | Should -Not -BeNullOrEmpty