+++ src/Tests/Unit/Public/Invoke-MistralAIModel.Tests.ps1
@@ -356,7 +356,7 @@ InModuleScope 'pwshBedrock' {
                 Mock -CommandName Invoke-BDRRModel {
                     $response
                     $Region         | Should -BeExactly 'us-west-2'
-                    $ModelID        | Should -BeExactly 'mistral.mixtral-8x7b-instruct-v0:1'
+                    $ModelID        | Should -BeExactly 'us.mistral.mixtral-8x7b-instruct-v0:1'
                     $AccessKey      | Should -BeExactly 'ak'
                     $SecretKey      | Should -BeExactly 'sk'
                     $ContentType    | Should -BeExactly 'application/json'
@@ -379,7 +379,7 @@ InModuleScope 'pwshBedrock' {
                 Mock -CommandName Invoke-BDRRModel {
                     $response
                     $Region             | Should -BeExactly 'us-west-2'
-                    $ModelID            | Should -BeExactly 'mistral.mixtral-8x7b-instruct-v0:1'
+                    $ModelID            | Should -BeExactly 'us.mistral.mixtral-8x7b-instruct-v0:1'
                     $Credential         | Should -Not -BeNullOrEmpty
                     $EndpointUrl        | Should -BeExactly 'string'
                     $NetworkCredential  | Should -Not -BeNullOrEmpty