+++ src/pwshBedrock/Public/Invoke-AI21LabsJambaModel.ps1
@@ -257,9 +257,11 @@ function Invoke-AI21LabsJambaModel {
     $jsonBody = $bodyObj | ConvertTo-Json -Depth 10
     [byte[]]$byteArray = [System.Text.Encoding]::UTF8.GetBytes($jsonBody)
 
+    $inferenceModelID = Format-InferenceProfileID -ModelID $ModelID -Region $Region
+
     $cmdletParams = @{
         ContentType = 'application/json'
-        ModelId     = $ModelID
+        ModelId     = $inferenceModelID
         Body        = $byteArray
     }