+++ docs/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.2.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.40.0]
+## [0.42.0]
 
 - Module changes:
     - Added two new properties to all model documentation objects:
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
         - `Video` - designates if the model can generate video
         - `Vision` - re-purposed to now designate if the model can see and describe uploaded content
             - This required some logic re-work in several functions that were previously using `Vision` to designate image capability
+    - Added inference profile ID support to all models IAW [Supported cross-region inference profiles](https://docs.aws.amazon.com/bedrock/latest/userguide/inference-profiles-support.html)
+        - pwshBedrock will now properly format the inference profile ID for all models that support which enables the ability to make Bedrock calls to other regions
     - Amazon
         - `Invoke-AmazonNovaTextModel` - added support for new Nova Text models: `amazon.nova-pro-v1:0`, `amazon.nova-lite-v1:0`, `amazon.nova-micro-v1:0`